Move v1-specific training_utils to training_utils_v1.py
PiperOrigin-RevId: 333612671
