+++ tensorflow/python/keras/engine/training_generator_v1.py
@@ -31,7 +31,6 @@ from tensorflow.python.framework import errors
 from tensorflow.python.keras import backend
 from tensorflow.python.keras import callbacks as cbks
 from tensorflow.python.keras.engine import training_utils
-from tensorflow.python.keras.engine import training_utils_v1
 from tensorflow.python.keras.utils import data_utils
 from tensorflow.python.keras.utils import generic_utils
 from tensorflow.python.keras.utils.mode_keys import ModeKeys
@@ -133,7 +132,7 @@ def model_iteration(model,
     original_dataset = data
     if steps_per_epoch is None:
       reset_dataset_after_each_epoch = True
-      steps_per_epoch = training_utils_v1.infer_steps_for_dataset(
+      steps_per_epoch = training_utils.infer_steps_for_dataset(
           model, data, steps_per_epoch, epochs=epochs, steps_name=steps_name)
 
   # Convert to a format that supports `next(generator)`.
@@ -180,11 +179,9 @@ def model_iteration(model,
       mode=mode)
 
   if mode == ModeKeys.PREDICT:
-    aggregator = training_utils_v1.OutputsAggregator(
-        True, steps=steps_per_epoch)
+    aggregator = training_utils.OutputsAggregator(True, steps=steps_per_epoch)
   else:
-    aggregator = training_utils_v1.MetricsAggregator(
-        True, steps=steps_per_epoch)
+    aggregator = training_utils.MetricsAggregator(True, steps=steps_per_epoch)
 
   should_set_learning_phase = context.executing_eagerly() and model.run_eagerly
   if should_set_learning_phase:
@@ -296,7 +293,7 @@ def model_iteration(model,
 
     # Run the test loop every epoch during training.
     if (do_validation and
-        training_utils_v1.should_run_validation(validation_freq, epoch) and
+        training_utils.should_run_validation(validation_freq, epoch) and
         not callbacks.model.stop_training):
       val_results = model_iteration(
           model,
@@ -541,7 +538,7 @@ def _get_num_samples_or_steps(data, steps_per_epoch):
   return steps_per_epoch, True
 
 
-class GeneratorOrSequenceTrainingLoop(training_utils_v1.TrainingLoop):
+class GeneratorOrSequenceTrainingLoop(training_utils.TrainingLoop):
   """Generator-like.
 
   Input is Python generator, or Sequence object.
@@ -572,7 +569,7 @@ class GeneratorOrSequenceTrainingLoop(training_utils_v1.TrainingLoop):
           workers=1,
           use_multiprocessing=False):
     model._validate_or_infer_batch_size(batch_size, steps_per_epoch, x)
-    training_utils_v1.check_generator_arguments(
+    training_utils.check_generator_arguments(
         y, sample_weight, validation_split=validation_split)
     return fit_generator(
         model,
@@ -605,7 +602,7 @@ class GeneratorOrSequenceTrainingLoop(training_utils_v1.TrainingLoop):
                workers=1,
                use_multiprocessing=False):
     model._validate_or_infer_batch_size(batch_size, steps, x)
-    training_utils_v1.check_generator_arguments(y, sample_weight)
+    training_utils.check_generator_arguments(y, sample_weight)
     return evaluate_generator(
         model,
         x,
@@ -638,7 +635,7 @@ class GeneratorOrSequenceTrainingLoop(training_utils_v1.TrainingLoop):
         use_multiprocessing=use_multiprocessing)
 
 
-class EagerDatasetOrIteratorTrainingLoop(training_utils_v1.TrainingLoop):
+class EagerDatasetOrIteratorTrainingLoop(training_utils.TrainingLoop):
   """A non-distributed Dataset or iterator in eager execution."""
 
   def fit(self,
@@ -661,11 +658,10 @@ class EagerDatasetOrIteratorTrainingLoop(training_utils_v1.TrainingLoop):
           **kwargs):
     model._validate_or_infer_batch_size(batch_size, steps_per_epoch, x)
     # Make sure that y, sample_weights, validation_split are not passed.
-    training_utils_v1.validate_dataset_input(x, y, sample_weight,
-                                             validation_split)
+    training_utils.validate_dataset_input(x, y, sample_weight, validation_split)
     if (isinstance(x, (dataset_ops.DatasetV1, dataset_ops.DatasetV2)) and
         shuffle):
-      training_utils_v1.verify_dataset_shuffled(x)
+      training_utils.verify_dataset_shuffled(x)
 
     return fit_generator(
         model,
@@ -695,7 +691,7 @@ class EagerDatasetOrIteratorTrainingLoop(training_utils_v1.TrainingLoop):
                **kwargs):
     model._validate_or_infer_batch_size(batch_size, steps, x)
     # Make sure that y, sample_weights, validation_split are not passed.
-    training_utils_v1.validate_dataset_input(x, y, sample_weight)
+    training_utils.validate_dataset_input(x, y, sample_weight)
     return evaluate_generator(
         model, x, steps=steps, verbose=verbose, workers=0, callbacks=callbacks)
 
@@ -712,7 +708,7 @@ class EagerDatasetOrIteratorTrainingLoop(training_utils_v1.TrainingLoop):
         model, x, steps=steps, verbose=verbose, workers=0, callbacks=callbacks)
 
 
-class GeneratorLikeTrainingLoop(training_utils_v1.TrainingLoop):
+class GeneratorLikeTrainingLoop(training_utils.TrainingLoop):
   """TrainingLoop that handle inputs like python generator.
 
   This is the default handler for most of the input data types, includes
@@ -759,9 +755,8 @@ class GeneratorLikeTrainingLoop(training_utils_v1.TrainingLoop):
                                                        validation_steps)
     elif validation_split and 0. < validation_split < 1.:
       (x, y, sample_weights, val_x, val_y,
-       val_sample_weights) = (
-           training_utils_v1.split_training_and_validation_data(
-               x, y, sample_weights, validation_split))
+       val_sample_weights) = training_utils.split_training_and_validation_data(
+           x, y, sample_weights, validation_split)
       validation_data = (val_x, val_y, val_sample_weights)
     else:
       if validation_steps:
